Adds ReadDriver for getCurrencies()
