+++ models/SQLite3Driver.js
@@ -1,6 +1,7 @@
 const sqlite3 = require('sqlite3').verbose();
 const fs = require('fs');
 const create = require('./SQLiteUtils/CreateDriver');
+const read = require('./SQLiteUtils/ReadDriver');
 
 function SQLite3Driver() {
     SQLite3Driver.prototype.dbName = './models/db/pixelshelf.db';
@@ -1152,28 +1153,10 @@ SQLite3Driver.prototype.getLibrarySize = function getLibrarySize() {
 
 SQLite3Driver.prototype.getCurrencies = function getCurrencies() {
     return new Promise(function (resolve, reject) {
-        SQLite3Driver.prototype.db = new sqlite3.Database(SQLite3Driver.prototype.dbName, sqlite3.OPEN_READONLY, (err) => {
-            if (err) {
-                reject(err);
-            }
-            let sql = `SELECT *
-                       FROM currency`;
-            SQLite3Driver.prototype.db.all(sql, [], (err, rows) => {
-                if (err) {
-                    reject(err);
-                } else {
-                    let result = [];
-                    rows.forEach((row) => {
-                        result.push({
-                            "code": row.code,
-                            "label": row.label,
-                            "symbol": row.symbol
-                        });
-                    });
-
-                    resolve(result);
-                }
-            });
+        read.selectCurrencies().then(res => {
+            resolve(res);
+        }).catch(err => {
+            reject(err);
         });
     });
 }
